Variable type bug fixed.

+++ MicroFrameworkPK_v4_3/DeviceCode/pal/Samraksh/Algorithms/Unwrap.cpp
@@ -63,7 +63,7 @@ BOOL calculatePhase(INT16* bufferI, INT16* bufferQ, UINT16* bufferUnwrap, INT32 
 {
 	int i;
 	int unwrappedPhase;
-	UINT16 minPhase, maxPhase;
+	int minPhase, maxPhase;
 	BOOL detection = false, threshholdMet = false;
 
 	for (i=0; i<length; i++){
